Rename 'PHP_VERSION' to 'BUILD_VERSION' and refactor Github workflow files
This commit renames the variable 'PHP_VERSION' to 'BUILD_VERSION' in Dockerfile and Github workflow files. This more accurately represents its purpose as it is not solely used for PHP versions. Furthermore, we have renamed the Github workflow file from 'build-image.yml' to 'build.yml' for simplicity and to better express its functionality.